Update and enable Poiseidon
## Description 

Enable the Poseidon hash function on mainnet.

There is a small change made to the function as well: It now only
accepts between 1 and 16 inputs. Previously, larger inputs were hashed
using an 16-ary Merkle tree which is compatible with zkLogin but not
other Poseiodon implementations which typically uses a 2- or 4-ary
Merkle tree.

This new version follows what seems to be the consensus on how to
implement Poseidon over BN254, and the limit of 16 inputs in the new
version is compatible with
https://github.com/iden3/go-iden3-crypto/tree/master/poseidon,
https://github.com/iden3/circomlib/blob/master/circuits/poseidon_constants.circom
and https://github.com/arnaucube/poseidon-rs. We leave it to the user to
decide how to handle larger inputs but provide an example (in form of a
unit test) of how to it for an 16-ary Merkle tree as used in zkLogin.

See also the SIP,
https://forums.haneul.io/t/sip-12-poseidon-in-haneul-move/45136.

## Test plan 

Unit tests.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [x] Protocol: Updates the protocol version and enables the Poseidon
hash function.
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: