cli(rpc): start consistent store separately
## Description

Guard the starting of the consistent store behind its own flag:
--with-consistent-store, that works similarly to --with-indexer in that
it will automatically be turned on if --with-graphql is passed to the
CLI.

The consistent store that is started will persist as well -- being
stored in the config directory that gets set-up for configs and node
storage.

## Test plan

Manually tested.