[static-ptbs] Charge for size of linkages during translation
## Description 

Add logic to charge for the size of the number of packages present in a
linkage after loading a PTB. Note that this is for the total set of
packages, and not just for the final linkage (in particular, we charge
per-package, per-version, and not just for per-package that is present
post linkage resolution).

## Test plan 

CI. 

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: