[haneul-adapter] Queue up gas charge change for `ReadRef`
## Description 

Reweight the charge for `ReadRef`. This is gated behind the existing
reweighting configuration.

Note: This is currently based on the gas model version -- the PR above
this will switch this over to be directly based on the protocol version,
but there's some plumbing involved that I am going to do in that PR.

## Test plan 

CI

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: