indexer-alt: updates to
## Description

An alternate way to implement inducing a final watermark commit before
shutdown.

The original implementation factored out the logic for gathering
watermarks but introduced duplicate code for establishing a connection
and writing out the data which was similar to but not the same as the
code used in the main loop.

This implementation moves the control-flow around so that when
cancellation is triggered, the main loop body is invoked one more time,
to avoid any code duplication.

## Test plan

The original test in  was succeeding even if the implementation
didn't make an attempt to sync before shutting down. Using the
cancellation token to stop the process in the test can demonstrate a
failure case, and the test has then been further hardened to check that
all watermarks have been processed by the task before it is cancelled,
so we know there is something to commit.

```
$ cargo nextest run -p haneul-indexer-alt-framework -- sync_on_shutdown
```

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: