add a pre-push hook for rustfmt
## Description 

This change aims to reduce the incidence of developers pushing code that
fails in CI, both as a developer convenience and a cost reduction
strategy for CI.

## Test plan 

git push ran rustfmt when I pushed this branch for PR

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] Indexing Framework: