[move-ide] Cached AST and auto-completion info optimizations
## Description 

This PR includes the first set of additional optimizations of the
`move-analyzer`'s memory footprint. Memory usage profiling revealed that
cached ASTs and cached auto-completion information is the main culprit
here. This PR mitigates these problems in the following way:
1. Previously (and incorrectly), even though we cache computed symbols
for dependencies we were still caching dependency ASTs which was
completely unnecessary. This PR fixes this.
2. Previously we were caching `CompilerInfo` for future re-use (in
analysis) and stored it in `Symbols` (for auto-completion requests).
Furthermore, we were saving `CompilerInfo` for all files, which is
necessary for analysis but unnecessary for auto-completion
(symbolication is run synchronously before auto-completion is served and
only need auto-completion for the file where the request was issued).
This PR splits compiler into two parts, one for auto-completion (which
goes to `Symbols` but is not additionally cached) and the rest (which
does not go to `Symbols` anymore but is cached for re-use).
Additionally, auto-completion information in `Symbols` is filtered to
represent only the current file

On a multi-project large workspace, this reduces memory footprint from
around 6GB to 1GB. On smaller workspaces the gain is up to 3x memory
savings

## Test plan 

Testing harness needed to be updated to accommodate caching strategy
changes. It is now also better equipped to catch caching-related
problems. All existing tests must pass.