[tracing] Move assertion to handle errors in instructions properly in `close_instruction`
## Description 

We were not properly handling the case where an instruction errored
between pre- and post instruction states. This fixes the assertion to
only check in the cases where the instruction was able to be executed
and traced without error.

## Test plan 

Ran on a local repro and verified the fix.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] Indexing Framework: