Make TestTransactionBuilder composable
## Description 

Previously TestTransactionBuilder can only be used to do one thing at a
time. This makes it hard to compose and do multiple things in the same
PTB in tests.
This PR refactors TestTransactionBuilder so that all it holds now is a
single PTB builder, and all changes are done to it incrementally. This
means that now we can compose and add more actions to the same
transaction easily in tests.

## Test plan 

CI

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] Indexing Framework: