refactor(indexer-alt): CancellationToken -> Services
## Description 

Make use of the new `Services` abstraction to avoid handling
cancellation using `CancellationToken`, which often requires threading
the token deep into each service.

This change also improves graceful shutdown support (for example when
there is a metrics service running alongside the indexer, and the
indexer triggers a shutdown, the metrics service is allowed time to
drain connections before shutting down itself).

## Test plan 

CI

## Stack

-  
-  

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [x] Indexing Framework: The indexer, ingestion service, and metrics
service all now return a `Service` instead of a `JoinHandle<()>` when
run. Use `Service::main` to wait for the service to exit cleanly or with
an error, or respond to a termination signal with a graceful shutdown.
`Service` also exposes `wait_for_shutdown`, `join`, and `shutdown`
functions to customise various aspects of the shutdown process.