chore(reader): demote SystemPackageTask error! logs to warn!-s
## Description

We typically reserve `error!` for messages related to situations that
the system does not know how to recover from, but the `error!` logs in
the system package task are related to issues that it is attempting to
recover from through retries, so they should really be `warn!` logs.

## Test plan

:eyes:

## Stack

-  
-  
-  

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] Indexing Framework: