display: option auto-dereference
## Description

Treat options as transparent, so that formats don't need to know about
how options are represented to look inside them. This extends the
support that already existed to dereference an option at the end of a
accessor path.

This is to get around the fact that options are represented in a
slightly strange way in Move (as a vector with zero or one elements).

## Test plan

New unit tests:

```
$ cargo nextest run -p haneul-display
```