refactor(display): factor formatting out of interpreter
## Description

Remove logic related to formatting/writing from the interpreter, so that
the interpreter can focus on converting ASTs into Values. This is in
preparation for implementing `MoveValue.extract` in GraphQL, which
requires evaluating a single Display expression against a Move Value.
Previously the interpreter also needed to be configured with limits
related to outputting, and now it does not.

## Test plan

Existing unit tests:

```
$ cargo nextest run -p haneul-display
```