haneul-indexer-alt: improve error message for pg/kv eventual consistency error
## Description 

Before this PR, the error message does not distinguish between records
not found in Postgres and records found in Postgres, but not KV:
```
2025-12-16T16:30:47.857516Z ERROR haneul_indexer_alt_jsonrpc::metrics::middleware: Request failed with internal error method="haneulx_getLatestHaneulSystemState" params="[]" code=-32603 response="{\"jsonrpc\":\"2.0\",\"id\":379775,\"error\":{\"code\":-32603,\"message\":\"Internal Error: Failed to fetch system state wrapper object: No data found\"}}" elapsed_ms=1006.9102059999999 
```

This PR improves the error message when the record is found in Postgres,
but not KV.

## Test plan 

This code is currently untested until kv simulation is added.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] Indexing Framework: