Add missing call index column to move calls table
## Description

We noticed the move call rows aren't unique due to repeated calls to the
same function in the same txn. This is causing query problems both for
dune and internally. Adding a call idx column to make this unique.

I wont merge this to main until the schema migration is complete.