graphql: MoveValue.asAddress
## Description

Support interpreting a `MoveValue` as an address if it has the
appropriate type. This allows for writing nested queries that traverse
wrapped objects as well as ID and address pointers.

This also required introducing another new field: `Address.addressAt`,
which works similarly to `objectAt`, but for addresses (and only
offering a way to re-scope it to a new checkpoint). This is to allow
someone to write a query that fetches a wrapped object but then needs to
run an owned object query on it (which is only valid if queried w.r.t. a
checkpoint, not a root version bound which is the usual scope we apply
to a wrapped object).

## Test plan

New E2E tests:

```
haneul-indexer-alt-e2e-tests$ cargo nextest run -- as_address
```