graphql: raise 500 for panics
## Description

Add a `CatchPanicLayer` to the GraphQL service. If the service is
compiled with panic unwinding enabled, this will catch panics during
request handling and translate them into a GraphQL Internal Server Error
response.

This change also introduces a `release-unwind` profile that the RPC
services can be compiled with (our default release profile aborts on
panic).

## Test plan

Insert a panic in a field resolver, query that field, and note that the
service returns a 500 and does not go down itself.