graphql: allow seeking ahead of checkpoint viewed at
## Description

Make it possible to reset "checkpoint viewed at" above the current
"checkpoint viewed at" as long as it's at or below the service's
checkpoint high watermark.

Previously a checkpoint bound was only allowed to travel further back in
time. This was a problem for values that used a checkpoint bound instead
of a root version bound (as is the case for the system state value, or
config objects), as it was not possible to query the latest version of
an object from one of these bounded values.

Now, checkpoint bounds are allowed to extend up to the latest known
checkpoint to the service, and `*At` queries (`addressAt`, `objectAt`,
`packageAt`) default to fetching the latest version of the entity known
to the service (not the latest as of the checkpoint being viewed).

The equivalent `Query` APIs (`Query.address`, `Query.object`,
`Query.package`) still behave as they did before, defaulting to fetching
the entity at the viewed checkpoint.

As part of this change, `addressAt` has been bolstered to work more
in line with `objectAt` and `packageAt` in that both `Query.address` and
`IAddressable.addressAt` support the same set of bounds, and there is a
multi-get API introduced.

## Test plan

E2E tests:

```
$ cargo nextest run -p haneul-indexer-alt-e2e-tests -- \
  graphql/at                                        \
  graphql/epochs/validator_set/exchange_rates
```